Fix void credit receipts showing up in undeposited list
Happens as long as the source receipt is not marked as void. This is because the logic was incorrectly set to `for_receipt_reference_id` instead of `receipt_reference_id`.

Setting to `for_receipt_reference_id` will only affect source receipts. Making credit receipts appear in undeposited list even if they are marked as void, as long as the source receipt is not marked as void.