Disable `Create Accounting System` button on load
Only enable button when loading is finished. If the user clicks the button when the page hasn't fully loaded results to an error as the client wasn't able to get the subscription information.

The only exception of this is when the subscription already reached its limit. Though this was addressed by using the `data-disabled` attribute. By default, the page loads and the button is disabled. Once the page has fully loaded, it checks for its limit status through the `data-disabled` attribute. If it is present, then remain `disabled`. Otherwise, enable button.